requirements.txt
# Core dependencies
GitPython>=3.1.40

# pygit2 often requires native toolchain; keep it off macOS by default.
pygit2>=1.14.1; platform_system != "Darwin"

# Text processing and tokenization
tokenizers>=0.15.0
transformers>=4.36.2

# Model training and inference
# torch wheels exist for macOS/Linux; on macOS this enables MPS.
torch>=2.5.0
pytorch-lightning>=2.1.3

# Parameter-Efficient Fine-Tuning (LoRA)
peft>=0.7.0

# Quantization support (4-bit, 8-bit)
# bitsandbytes is CUDA-only and does not provide macOS wheels.
bitsandbytes>=0.41.0; platform_system != "Darwin"

# Data processing
datasets>=2.16.1
huggingface_hub>=0.20.1

# Numerical computation
numpy>=1.24.3
pandas>=2.0.0
scikit-learn>=1.3.0

# Vector embeddings and search
sentence-transformers>=2.2.0
qdrant-client>=1.7.0

# Utilities
tqdm>=4.66.0
pydantic>=2.0.0
python-dotenv>=1.0.0
click>=8.0.0
rich>=13.0.0
pyyaml>=6.0

# System monitoring
psutil>=5.9.0

# Serialization
msgpack>=1.0.7

# Optional: For better performance
accelerate>=0.24.0  # Multi-GPU/distributed training support

# Optional test dependency
pytest>=7.4.0
